move spring-boot logic to spring-boot generator
