+++ generators/spring-boot/generator.spec.ts
@@ -1,13 +1,12 @@
 import { basename, dirname } from 'path';
 import { fileURLToPath } from 'url';
 import { before, it, describe, expect } from 'esmocha';
-import lodash from 'lodash';
-
+import { snakeCase } from 'lodash-es';
+import { checkEnforcements, defaultHelpers as helpers, runResult } from '../../test/support/index.js';
 import { shouldSupportFeatures, testBlueprintSupport } from '../../test/support/tests.js';
-import Generator from './index.js';
-import { defaultHelpers as helpers, result } from '../../test/support/index.js';
+import Generator from '../server/index.js';
 
-const { snakeCase } = lodash;
+import { mockedGenerators } from '../server/__test-support/index.js';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
@@ -20,14 +19,25 @@ describe(`generator - ${generator}`, () => {
   });
   shouldSupportFeatures(Generator);
   describe('blueprint support', () => testBlueprintSupport(generator));
+  checkEnforcements({}, generator);
 
-  describe('with default config', () => {
+  describe('with jwt', () => {
     before(async () => {
-      await helpers.runJHipster(generator).withJHipsterConfig({ skipClient: true }).withMockedSource();
+      await helpers.runJHipster(generator).withJHipsterConfig({ authenticationType: 'jwt' }).withMockedGenerators(mockedGenerators);
     });
 
-    it('should match files snapshot', () => {
-      expect(result.getStateSnapshot()).toMatchSnapshot();
+    it('should match generated files snapshot', () => {
+      expect(runResult.getStateSnapshot()).toMatchSnapshot();
+    });
+  });
+
+  describe('with oauth2', () => {
+    before(async () => {
+      await helpers.runJHipster(generator).withJHipsterConfig({ authenticationType: 'oauth2' }).withMockedGenerators(mockedGenerators);
+    });
+
+    it('should match generated files snapshot', () => {
+      expect(runResult.getStateSnapshot()).toMatchSnapshot();
     });
   });
 });
