+++ generators/spring-data-elasticsearch/generator.js
@@ -73,6 +73,12 @@ export default class ElasticsearchGenerator extends BaseApplicationGenerator {
 
   get postWriting() {
     return this.asPostWritingTaskGroup({
+      addTestSpringFactory({ source, application }) {
+        source.addTestSpringFactory({
+          key: 'org.springframework.test.context.ContextCustomizerFactory',
+          value: `${application.packageName}.config.TestContainersSpringContextCustomizerFactory`,
+        });
+      },
       addDependencies({ application, source }) {
         if (application.buildToolMaven) {
           source.addMavenProperty?.({
